Update contributing, deduplicate more functions
