+++ extensions/google-antigravity-auth/index.ts
@@ -1,8 +1,8 @@
 import { createHash, randomBytes } from "node:crypto";
-import { readFileSync } from "node:fs";
 import { createServer } from "node:http";
 import {
   emptyPluginConfigSchema,
+  isWSL2Sync,
   type OpenClawPluginApi,
   type ProviderAuthContext,
 } from "openclaw/plugin-sdk";
@@ -52,32 +52,8 @@ function generatePkce(): { verifier: string; challenge: string } {
   return { verifier, challenge };
 }
 
-function isWSL(): boolean {
-  if (process.platform !== "linux") {
-    return false;
-  }
-  try {
-    const release = readFileSync("/proc/version", "utf8").toLowerCase();
-    return release.includes("microsoft") || release.includes("wsl");
-  } catch {
-    return false;
-  }
-}
-
-function isWSL2(): boolean {
-  if (!isWSL()) {
-    return false;
-  }
-  try {
-    const version = readFileSync("/proc/version", "utf8").toLowerCase();
-    return version.includes("wsl2") || version.includes("microsoft-standard");
-  } catch {
-    return false;
-  }
-}
-
 function shouldUseManualOAuthFlow(isRemote: boolean): boolean {
-  return isRemote || isWSL2();
+  return isRemote || isWSL2Sync();
 }
 
 function buildAuthUrl(params: { challenge: string; state: string }): string {
